Fix bugs found during real-hardware smoke test on the GTX 1080 Ti desktop
- cuda_backend: default compute_type float16 -> float32. Pascal GPUs
  (GTX 1080 Ti) don't support efficient float16 in ctranslate2; it
  raised on the very machine this backend targets.
- diarizer: load audio via soundfile into an in-memory waveform dict
  instead of handing pyannote a path. Path-based decoding goes through
  torchcodec, whose native libs must match the installed FFmpeg AND
  torch build exactly - this broke twice across unrelated env changes.
  Also unwrap pyannote.audio>=4's DiarizeOutput (exclusive_speaker_diarization)
  instead of assuming a bare Annotation is returned.

Verified end-to-end against real Postgres, real CUDA ASR, real
diarization (HF gated models), and real Groq API on the GTX 1080 Ti
machine: capture format -> transcribe -> diarize -> merge -> summarize
-> docx export -> DB rows, all succeeded without crashing.

+++ app/asr/cuda_backend.py
@@ -6,7 +6,10 @@ from app.asr.base import TranscriptSegmentResult
 
 
 class FasterWhisperBackend:
-    def __init__(self, model_size: str = "large-v3", device: str = "cuda", compute_type: str = "float16"):
+    def __init__(self, model_size: str = "large-v3", device: str = "cuda", compute_type: str = "float32"):
+        # float16 isn't "efficient" on Pascal-generation GPUs (e.g. GTX 1080 Ti,
+        # one of the two mandated target machines) and ctranslate2 raises rather
+        # than silently falling back; float32 works everywhere.
         self._model = WhisperModel(model_size, device=device, compute_type=compute_type)
 
     def transcribe(self, wav_path: Path, language: str = "id") -> list[TranscriptSegmentResult]:
